+++ CHANGELOG.md
@@ -1,16 +1,21 @@
-## CHANGELOG
+## [1.1.0] - 2022-11-12
 
----
+- Added new website to get proxies from: <https://free-proxy-list.net>
+- Added new website to get proxies from: <https://free-proxy-list.net/uk-proxy.html>
+- Added new website to get proxies from: <https://www.us-proxy.org/>
+- Change lxml version to 4.9.1
 
 ## [1.0.6] - 2022-01-23
 
 - Added `google` parameter
 - Added `https` parameter
+
 ## [1.0.5] - 2022-01-07
 
 - Added `elite` parameter
 - Add exception class and raise exception instead of system exit
 - Change lxml version to 4.6.5
+
 ## [1.0.4] - 2021-11-13
 
 - Fix proxy list default length
@@ -31,3 +36,5 @@
 - Fix imports
 
 ## [1.0.0] - 2019-02-04
+
+- Initial release